Change Double to Float Cannot start an expression with "×" or "÷"
Main.storyboard:
Disable "×" and "÷" buttons

ViewController.swift:
Create
	priorityOperatorButtons IBOutlet collection
	haveToDisablePriorityOperatorButtons computed property
	switchPriorityOperatorButtonsEnableState()

CalculatorImplementation.swift:
Create
	assignValueForEachPartOfExpression()
	canPerformCalculation()
	textToComputeHasRelativeSign computed property
Refactor calculate() with assignValueForEachPartOfExpression() and canPerformCalculation()
Remove
	canDivide computed property
	canAddOperator computed property
Modify
	expressionIsCorrect computed property with multiply and divide symbols
	add(mathOperator:) with textToComputeHasRelativeSign

SimpleCalcTests.swift:
Add test to check replace relative sign with another

ErrorMessage.swift:
Implements String to use rawValue
Remove name property

+++ CountOnMe/Model/Calculation/Calculator.swift
@@ -13,5 +13,5 @@ protocol Calculator {
 
     func add(number: Int)
     func add(mathOperator: MathOperator)
-    func calculate() -> Double?
+    func calculate() -> Float?
 }
